Refactor: habit 정렬 유틸리티 함수로 분리 #92
