+++ src/components/mypage/habits/HabitList.tsx
@@ -4,10 +4,7 @@ import HabitItem from './HabitItem';
 import { HabitFormData, HabitWithPoints } from '@/types/habits.type';
 import { useCreateHabitMutation } from '@/lib/mutations/useHabitMutation';
 import { useMemo } from 'react';
-import {
-  getCooldownStatus,
-  getCurrentDayStatus,
-} from '@/lib/utils/habit.utils';
+import { sortHabitsByEnabled } from '@/lib/utils/habit.utils';
 
 type HabitListProps = {
   userId: string;
@@ -36,15 +33,7 @@ const HabitList = ({
   };
 
   // 쿨다운 없음, 오늘 수행 가능 -> 위쪽에 정렬
-  const sortedHabits = useMemo(() => {
-    const isEnabled = (habit: HabitWithPoints) =>
-      !getCooldownStatus(habit.userPoints) && getCurrentDayStatus(habit);
-    return [...habits].sort((a, b) => {
-      const aEnabled = isEnabled(a);
-      const bEnabled = isEnabled(b);
-      return aEnabled === bEnabled ? 0 : aEnabled ? -1 : 1;
-    });
-  }, [habits]);
+  const sortedHabits = useMemo(() => sortHabitsByEnabled(habits), [habits]);
 
   return (
     <ul className="h-[460px] overflow-y-auto">
